# Pertemuan 11/Restfull API/public/assets/api.yaml
openapi: 3.0.0
paths:
...
  /provinces:
    get:
      tags:
        - Provinces
      summary: 'Find list Provinces'
      description: 'Returns list of Provinces'
      operationId: getProvinces
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Provinces' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Provinces' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/Provinces' } }
                type: object
        '404':
          description: 'Provinces not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - Provinces
      summary: 'Add a new Provinces to the store'
      operationId: addProvinces
      requestBody:
        $ref: '#/components/requestBodies/Provinces'
      responses:
        '201':
          description: 'Created Provinces'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provinces'
            application/xml:
              schema:
                $ref: '#/components/schemas/Provinces'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/provinces/{id}':
    get:
      tags:
        - Provinces
      summary: 'Find Provinces by ID'
      description: 'Returns a single Provinces'
      operationId: getProvincesById
      parameters:
        -
          name: id
          in: path
          description: 'ID of Provinces to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provinces'
            application/xml:
              schema:
                $ref: '#/components/schemas/Provinces'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'Provinces not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - Provinces
      summary: 'Update an existing Provinces'
      operationId: updateProvinces
      parameters:
        -
          name: id
          in: path
          description: 'Provinces id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/Provinces'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Provinces not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - Provinces
      summary: 'Deletes a Provinces'
      operationId: deleteProvinces
      parameters:
        -
          name: id
          in: path
          description: 'Provinces id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
  /publicPollutions:
    get:
      tags:
        - PublicPollutions
      summary: 'Find list PublicPollutions'
      description: 'Returns list of PublicPollutions'
      operationId: getPublicPollutions
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/PublicPollutions' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/PublicPollutions' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/PublicPollutions' } }
                type: object
        '404':
          description: 'PublicPollutions not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - PublicPollutions
      summary: 'Add a new PublicPollutions to the store'
      operationId: addPublicPollutions
      requestBody:
        $ref: '#/components/requestBodies/PublicPollutions'
      responses:
        '201':
          description: 'Created PublicPollutions'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicPollutions'
            application/xml:
              schema:
                $ref: '#/components/schemas/PublicPollutions'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/publicPollutions/{id}':
    get:
      tags:
        - PublicPollutions
      summary: 'Find PublicPollutions by ID'
      description: 'Returns a single PublicPollutions'
      operationId: getPublicPollutionsById
      parameters:
        -
          name: id
          in: path
          description: 'ID of PublicPollutions to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicPollutions'
            application/xml:
              schema:
                $ref: '#/components/schemas/PublicPollutions'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'PublicPollutions not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - PublicPollutions
      summary: 'Update an existing PublicPollutions'
      operationId: updatePublicPollutions
      parameters:
        -
          name: id
          in: path
          description: 'PublicPollutions id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/PublicPollutions'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'PublicPollutions not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - PublicPollutions
      summary: 'Deletes a PublicPollutions'
      operationId: deletePublicPollutions
      parameters:
        -
          name: id
          in: path
          description: 'PublicPollutions id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
...
components:
  schemas:
...
    Provinces:
      title: Provinces
      description: Provinces
      properties:
        prov_id:
          nullable: false
          title: prov_id
          description: prov_id
          type: integer
          format: '-'
        prov_name:
          nullable: true
          title: prov_name
          description: prov_name
          type: string
          format: '-'
          maxLength: 255
        locationid:
          nullable: true
          title: locationid
          description: locationid
          type: integer
          format: '-'
        status:
          nullable: true
          title: status
          description: status
          type: integer
          format: '-'
      type: object
    PublicPollutions:
      title: PublicPollutions
      description: PublicPollutions
      properties:
        public_pollution_id:
          nullable: false
          title: public_pollution_id
          description: public_pollution_id
          type: string
          format: '-'
          maxLength: 6
        observer_id:
          nullable: false
          title: observer_id
          description: observer_id
          type: string
          format: '-'
          maxLength: 6
        public_pollution_title:
          nullable: false
          title: public_pollution_title
          description: public_pollution_title
          type: string
          format: '-'
          maxLength: 70
        public_pollution_description:
          nullable: false
          title: public_pollution_description
          description: public_pollution_description
          type: string
          format: '-'
        public_pollution_longitude:
          nullable: false
          title: public_pollution_longitude
          description: public_pollution_longitude
          type: string
          format: '-'
          maxLength: 100
        public_pollution_lattitude:
          nullable: false
          title: public_pollution_lattitude
          description: public_pollution_lattitude
          type: string
          format: '-'
          maxLength: 100
        public_pollution_type:
          nullable: false
          title: public_pollution_type
          description: public_pollution_type
          type: string
          format: '-'
        public_pollution_img:
          nullable: false
          title: public_pollution_img
          description: public_pollution_img
          type: string
          format: '-'
          maxLength: 255
        public_pollution_status:
          nullable: false
          title: public_pollution_status
          description: public_pollution_status
          type: string
          format: '-'
        public_pollution_action:
          nullable: false
          title: public_pollution_action
          description: public_pollution_action
          type: string
          format: '-'
        public_pollution_action_img:
          nullable: false
          title: public_pollution_action_img
          description: public_pollution_action_img
          type: string
          format: '-'
          maxLength: 255
        created_at:
          nullable: false
          title: created_at
          description: created_at
          type: string
          format: '-'
        updated_at:
          nullable: false
          title: updated_at
          description: updated_at
          type: string
          format: '-'
      type: object
...
  requestBodies:
...
    Provinces:
      description: 'Provinces object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/Provinces'
        application/xml:
          schema:
            $ref: '#/components/schemas/Provinces'
        application/json:
          schema:
            $ref: '#/components/schemas/Provinces'
    PublicPollutions:
      description: 'PublicPollutions object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/PublicPollutions'
        application/xml:
          schema:
            $ref: '#/components/schemas/PublicPollutions'
        application/json:
          schema:
            $ref: '#/components/schemas/PublicPollutions'
...